+++ resources/themes/pterodactyl/base/index.blade.php
@@ -78,6 +78,11 @@
                     </tbody>
                 </table>
             </div>
+            @if($servers->hasPages())
+                <div class="box-footer">
+                    <div class="col-md-12 text-center">{!! $servers->render() !!}</div>
+                </div>
+            @endif
         </div>
     </div>
 </div>